+++ changelog.md
@@ -1,3 +1,6 @@
+# v1.0.4
+- Minor tweaks
+
 # v1.0.3
 - Code cleanup
 - Fixed swipe mode triggering while touching menu [*Windows*, *Android*]